allow launch JS component to read env variables from env.json
