+++ public/index.html
@@ -9,7 +9,6 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
     <title>Washington Opioid Clinical Assessment</title>
-    <script src="js/index.js"></script>
   </head>
   <body>
     <noscript>
